+++ android/src/main/java/trade/ksanbal/esp_blufi_for_flutter/params/BlufiConfigureParams.java
@@ -1,4 +1,4 @@
-package vn.vconnex.blufi_plugin.params;
+package trade.ksanbal.esp_blufi_for_flutter.params;
 
 import java.io.Serializable;
 import java.util.Locale;